refactored register and login endpoints to make use of db controller for cleaner routes list
