+++ server/index.js
@@ -3,8 +3,7 @@ const express = require('express');
 const path = require('path');
 const app = express();
 const port = 8080;
-const bcrypt = require('bcrypt');
-const jwt = require('jsonwebtoken');
+const { register, login } = require('../database/controllers/authentication.js');
 //const goodbye = require('./routes/goodbye.js');
 
 // connect to db
@@ -27,81 +26,10 @@ app.get('/goodbye', (req, res) => {
 // ---- Authentication Routes ---- //
 
 // Register Endpoint 
-app.post('/register', (req, res) => {
-  bcrypt
-    .hash(req.query.password, 10)
-    .then((hashedPassword) => {
-      const user = new User({
-        full_name: req.query.name,
-        email: req.query.email,
-        password: hashedPassword,
-      });
-
-      user
-        .save()
-        .then((result) => {
-          res.status(201).send({
-            message: 'User Created Successfully',
-            result: result,
-          });
-        })
-        .catch((error) => {
-          res.status(500).send({
-            message: 'Error Creating User',
-            error: error,
-          });
-        });
-    })
-    .catch((error) => {
-      res.status(500).send({
-        message: 'Password was not hashed successfully',
-        error: error,
-      });
-    });
-});
+app.post('/register', register);
 
 // Login Endpoint
-app.post("/login", (req, res) => {
-  User.findOne({ email: req.query.email })
-    .then((user) => {
-      bcrypt
-        .compare(req.query.password, user.password)
-        .then((passwordCheck) => {
-          if(!passwordCheck) {
-            return res.status(400).send({
-              message: "Password does not match",
-            });
-          }
-
-          const token = jwt.sign(
-            {
-              userId: user._id,
-              userEmail: user.email,
-            },
-            'RANDOM-TOKEN',
-            { expiresIn: '24h' }
-          );
-
-          res.status(200).send({
-            message: 'Login Successful',
-            email: user.email,
-            token,
-          });
-        })
-        .catch((error) => {
-          res.status(400).send({
-            message: 'Password does not match',
-            error,
-          });
-        });
-    })
-    .catch((error) => {
-      res.status(404).send({
-        message: 'Email not found',
-        error: error,
-      });
-    });
-});
+app.post("/login", login);
 
 // set port and listen for requests
